treewide: consolidate platform specific code in platform.c
Get rid of most __APPLE__ guards by introducing a central platform.c unit
providing drop-in replacements for missing APIs.

Also move system signal definitions into the new platform file to be able
to share them with the upcoming debug library.
