+++ CMakeLists.txt
@@ -50,7 +50,7 @@ INCLUDE(FindPkgConfig)
 PKG_CHECK_MODULES(JSONC REQUIRED json-c)
 INCLUDE_DIRECTORIES(${JSONC_INCLUDE_DIRS})
 
-SET(UCODE_SOURCES lexer.c lib.c vm.c chunk.c vallist.c compiler.c source.c types.c program.c)
+SET(UCODE_SOURCES lexer.c lib.c vm.c chunk.c vallist.c compiler.c source.c types.c program.c platform.c)
 ADD_LIBRARY(libucode SHARED ${UCODE_SOURCES})
 SET(SOVERSION 0 CACHE STRING "Override ucode library version")
 SET_TARGET_PROPERTIES(libucode PROPERTIES OUTPUT_NAME ucode SOVERSION ${SOVERSION})
