+++ types.c
@@ -17,7 +17,6 @@
 #include <stdarg.h>
 #include <stdlib.h>
 #include <assert.h>
-#include <endian.h>
 #include <errno.h>
 #include <math.h>
 #include <ctype.h>
